requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
firebase-admin==6.2.0
gcloud==0.17.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
jws==0.1.3
MarkupSafe==2.1.3
msgpack==1.0.5
oauth2client==3.0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
proto-plus==1.22.3
protobuf==4.24.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.8.0
pyparsing==3.1.1
python-jwt==2.0.1
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==0.7.0
rsa==4.9
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
ulid-py==1.1.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.6.0
wrapt==1.16.0
